+++ src/apis/record_auth_otp_request.test.ts
@@ -190,6 +190,9 @@ const scenarios: Scenario[] = [
         otp.SetCollectionRef(user.collection().Id);
         otp.SetRecordRef(user.Id);
         otp.ProxyRecord().SetPassword("123456");
+        const created = NowDateTime().Add(-1000 + i);
+        otp.ProxyRecord().SetRaw("created", created);
+        otp.ProxyRecord().SetRaw("updated", created);
         const err = await app.SaveNoValidate(otp);
         if (err) {
           throw new Error(err.message);